engine: isolate buildkit client+session to each client
This is a pre-req for upcoming changes to support passing host sockets
to modules and improvements to secret isolation.
