+++ cmd/introspect/main.go
@@ -4,16 +4,23 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/dagger/dagger/core"
 	"github.com/dagger/dagger/core/schema"
+	"github.com/dagger/dagger/dagql"
 )
 
 func main() {
 	ctx := context.Background()
-	srv, err := schema.New(ctx, schema.InitializeArgs{})
-	if err != nil {
+
+	root := &core.Query{}
+	dag := dagql.NewServer(root)
+	coreMod := &schema.CoreMod{Dag: dag}
+	coreModDeps := core.NewModDeps(root, []core.Mod{coreMod})
+	if err := coreMod.Install(ctx, dag); err != nil {
 		panic(err)
 	}
-	res, err := srv.Introspect(ctx)
+
+	res, err := coreModDeps.SchemaIntrospectionJSON(ctx, false)
 	if err != nil {
 		panic(err)
 	}
