+++ core/gqlformat_test.go
@@ -1,0 +1,55 @@
+package core
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestNamespaceObjects(t *testing.T) {
+	t.Parallel()
+
+	testCases := []struct {
+		testCase  string
+		namespace string
+		obj       string
+		result    string
+	}{
+		{
+			testCase:  "namespace",
+			namespace: "Foo",
+			obj:       "Bar",
+			result:    "FooBar",
+		},
+		{
+			testCase:  "namespace into camel case",
+			namespace: "foo",
+			obj:       "bar-baz",
+			result:    "FooBarBaz",
+		},
+		{
+			testCase:  "don't namespace when equal",
+			namespace: "foo",
+			obj:       "Foo",
+			result:    "Foo",
+		},
+		{
+			testCase:  "don't namespace when prefixed",
+			namespace: "foo",
+			obj:       "FooBar",
+			result:    "FooBar",
+		},
+		{
+			testCase:  "still namespace when prefixed if not full",
+			namespace: "foo",
+			obj:       "Foobar",
+			result:    "FooFoobar",
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.testCase, func(t *testing.T) {
+			result := namespaceObject(tc.obj, tc.namespace, tc.namespace)
+			require.Equal(t, tc.result, result)
+		})
+	}
+}
