+++ core/services.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/dagger/dagger/dagql/idproto"
 	"github.com/dagger/dagger/engine"
-	"github.com/dagger/dagger/engine/buildkit"
 	"github.com/dagger/dagger/network"
 	bkgw "github.com/moby/buildkit/frontend/gateway/client"
 	"github.com/moby/buildkit/util/bklog"
@@ -33,7 +32,6 @@ const (
 // Services manages the lifecycle of services, ensuring the same service only
 // runs once per client.
 type Services struct {
-	bk       *buildkit.Client
 	starting map[ServiceKey]*sync.WaitGroup
 	running  map[ServiceKey]*RunningService
 	bindings map[ServiceKey]int
@@ -77,9 +75,8 @@ type ServiceKey struct {
 }
 
 // NewServices returns a new Services.
-func NewServices(bk *buildkit.Client) *Services {
+func NewServices() *Services {
 	return &Services{
-		bk:       bk,
 		starting: map[ServiceKey]*sync.WaitGroup{},
 		running:  map[ServiceKey]*RunningService{},
 		bindings: map[ServiceKey]int{},
