+++ engine/server/buildkitcontroller.go
@@ -5,25 +5,17 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"runtime"
 	"runtime/debug"
 	"sync"
 	"time"
 
-	"github.com/dagger/dagger/auth"
-	"github.com/dagger/dagger/core"
-	"github.com/dagger/dagger/core/pipeline"
 	"github.com/dagger/dagger/engine"
-	"github.com/dagger/dagger/engine/buildkit"
-	"github.com/dagger/dagger/engine/cache"
 	controlapi "github.com/moby/buildkit/api/services/control"
 	apitypes "github.com/moby/buildkit/api/types"
 	"github.com/moby/buildkit/cache/remotecache"
 	bkclient "github.com/moby/buildkit/client"
 	"github.com/moby/buildkit/executor/oci"
 	"github.com/moby/buildkit/frontend"
-	bkgw "github.com/moby/buildkit/frontend/gateway/client"
-	"github.com/moby/buildkit/identity"
 	"github.com/moby/buildkit/session"
 	"github.com/moby/buildkit/session/grpchijack"
 	containerdsnapshot "github.com/moby/buildkit/snapshot/containerd"
@@ -181,10 +173,6 @@ func (e *BuildkitController) Session(stream controlapi.Control_SessionServer) (r
 		if !ok {
 			return fmt.Errorf("server %q not found", opts.ServerID)
 		}
-		err := srv.bkClient.VerifyClient(opts.ClientID, opts.ClientSecretToken)
-		if err != nil {
-			return fmt.Errorf("failed to verify client: %w", err)
-		}
 		bklog.G(ctx).Debugf("forwarding client to server")
 		err = srv.ServeClientConn(ctx, opts, conn)
 		if errors.Is(err, io.ErrClosedPipe) {
@@ -217,65 +205,10 @@ func (e *BuildkitController) Session(stream controlapi.Control_SessionServer) (r
 	if !ok {
 		bklog.G(ctx).Debugf("initializing new server")
 
-		getSessionCtx, getSessionCancel := context.WithTimeout(ctx, 10*time.Second)
-		defer getSessionCancel()
-		caller, err := e.SessionManager.Get(getSessionCtx, opts.ClientID, false)
+		srv, err = e.newDaggerServer(ctx, opts)
 		if err != nil {
 			e.perServerMu.Unlock(opts.ServerID)
-			return fmt.Errorf("get session: %w", err)
-		}
-		bklog.G(ctx).Debugf("connected new server session")
-
-		secretStore := core.NewSecretStore()
-		authProvider := auth.NewRegistryAuthProvider()
-
-		var cacheImporterCfgs []bkgw.CacheOptionsEntry
-		for _, cacheImportCfg := range opts.UpstreamCacheImportConfig {
-			_, ok := e.UpstreamCacheImporters[cacheImportCfg.Type]
-			if !ok {
-				e.perServerMu.Unlock(opts.ServerID)
-				return fmt.Errorf("unknown cache importer type %q", cacheImportCfg.Type)
-			}
-			cacheImporterCfgs = append(cacheImporterCfgs, bkgw.CacheOptionsEntry{
-				Type:  cacheImportCfg.Type,
-				Attrs: cacheImportCfg.Attrs,
-			})
-		}
-
-		// using a new random ID rather than server ID to squash any nefarious attempts to set
-		// a server id that has e.g. ../../.. or similar in it
-		progSockPath := fmt.Sprintf("/run/dagger/server-progrock-%s.sock", identity.NewID())
-
-		bkClient, err := buildkit.NewClient(ctx, buildkit.Opts{
-			Worker:                e.worker,
-			SessionManager:        e.SessionManager,
-			LLBSolver:             e.llbSolver,
-			GenericSolver:         e.genericSolver,
-			SecretStore:           secretStore,
-			AuthProvider:          authProvider,
-			PrivilegedExecEnabled: e.privilegedExecEnabled,
-			UpstreamCacheImports:  cacheImporterCfgs,
-			ProgSockPath:          progSockPath,
-			MainClientCaller:      caller,
-			MainClientCallerID:    opts.ClientID,
-			DNSConfig:             e.DNSConfig,
-			Frontends:             e.Frontends,
-		})
-		if err != nil {
-			e.perServerMu.Unlock(opts.ServerID)
-			return fmt.Errorf("new Buildkit client: %w", err)
-		}
-
-		bklog.G(ctx).Debugf("initialized new server buildkit client")
-
-		labels := opts.Labels
-		labels = append(labels, pipeline.EngineLabel(e.EngineName))
-		labels = append(labels, pipeline.LoadServerLabels(engine.Version, runtime.GOOS, runtime.GOARCH, e.cacheManager.ID() != cache.LocalCacheID)...)
-
-		srv, err = NewDaggerServer(ctx, bkClient, e.worker, caller, opts.ServerID, secretStore, authProvider, labels, opts.CloudToken, opts.DoNotTrack)
-		if err != nil {
-			e.perServerMu.Unlock(opts.ServerID)
-			return fmt.Errorf("new Dagger server: %w", err)
+			return fmt.Errorf("new APIServer: %w", err)
 		}
 		e.serverMu.Lock()
 		e.servers[opts.ServerID] = srv
@@ -290,12 +223,9 @@ func (e *BuildkitController) Session(stream controlapi.Control_SessionServer) (r
 			delete(e.servers, opts.ServerID)
 			e.serverMu.Unlock()
 
-			srv.Close()
-
-			if err := bkClient.Close(); err != nil {
-				bklog.G(ctx).WithError(err).Errorf("failed to close buildkit client for server %s", opts.ServerID)
+			if err := srv.Close(); err != nil {
+				bklog.G(ctx).WithError(err).Error("failed to close server")
 			}
-			bklog.G(ctx).Debug("closed buildkit client")
 
 			time.AfterFunc(time.Second, e.throttledGC)
 			bklog.G(ctx).Debug("server removed")
@@ -303,7 +233,7 @@ func (e *BuildkitController) Session(stream controlapi.Control_SessionServer) (r
 	}
 	e.perServerMu.Unlock(opts.ServerID)
 
-	err = srv.bkClient.RegisterClient(opts.ClientID, opts.ClientHostname, opts.ClientSecretToken)
+	err = srv.RegisterClient(opts.ClientID, opts.ClientHostname, opts.ClientSecretToken)
 	if err != nil {
 		return fmt.Errorf("failed to register client: %w", err)
 	}
@@ -325,52 +255,6 @@ func (e *BuildkitController) Session(stream controlapi.Control_SessionServer) (r
 		return fmt.Errorf("wait: %w", err)
 	}
 	return nil
-}
-
-// Solve is currently only used for triggering upstream remote cache exports on a dagger server
-func (e *BuildkitController) Solve(ctx context.Context, req *controlapi.SolveRequest) (*controlapi.SolveResponse, error) {
-	opts, err := engine.ClientMetadataFromContext(ctx)
-	if err != nil {
-		return nil, err
-	}
-	ctx = bklog.WithLogger(ctx, bklog.G(ctx).
-		WithField("client_id", opts.ClientID).
-		WithField("client_hostname", opts.ClientHostname).
-		WithField("server_id", opts.ServerID))
-
-	e.serverMu.RLock()
-	srv, ok := e.servers[opts.ServerID]
-	e.serverMu.RUnlock()
-	if !ok {
-		return nil, fmt.Errorf("unknown server id %q", opts.ServerID)
-	}
-	err = srv.bkClient.VerifyClient(opts.ClientID, opts.ClientSecretToken)
-	if err != nil {
-		return nil, fmt.Errorf("failed to register client: %w", err)
-	}
-
-	cacheExporterFuncs := make([]buildkit.ResolveCacheExporterFunc, len(req.Cache.Exports))
-	for i, cacheExportCfg := range req.Cache.Exports {
-		cacheExportCfg := cacheExportCfg
-		exporterFunc, ok := e.UpstreamCacheExporters[cacheExportCfg.Type]
-		if !ok {
-			return nil, fmt.Errorf("unknown cache exporter type %q", cacheExportCfg.Type)
-		}
-		cacheExporterFuncs[i] = func(ctx context.Context, sessionGroup session.Group) (remotecache.Exporter, error) {
-			return exporterFunc(ctx, sessionGroup, cacheExportCfg.Attrs)
-		}
-	}
-	if len(cacheExporterFuncs) > 0 {
-		// run cache export instead
-		bklog.G(ctx).Debugf("running cache export for client %s", opts.ClientID)
-		err := srv.bkClient.UpstreamCacheExport(ctx, cacheExporterFuncs)
-		if err != nil {
-			bklog.G(ctx).WithError(err).Errorf("error running cache export for client %s", opts.ClientID)
-			return &controlapi.SolveResponse{}, err
-		}
-		bklog.G(ctx).Debugf("done running cache export for client %s", opts.ClientID)
-	}
-	return &controlapi.SolveResponse{}, nil
 }
 
 func (e *BuildkitController) DiskUsage(ctx context.Context, r *controlapi.DiskUsageRequest) (*controlapi.DiskUsageResponse, error) {
@@ -547,6 +431,10 @@ func (e *BuildkitController) gc() {
 	if size > 0 {
 		bklog.G(ctx).Debugf("gc cleaned up %d bytes", size)
 	}
+}
+
+func (e *BuildkitController) Solve(ctx context.Context, req *controlapi.SolveRequest) (*controlapi.SolveResponse, error) {
+	return nil, fmt.Errorf("solve not implemented")
 }
 
 func (e *BuildkitController) Status(req *controlapi.StatusRequest, stream controlapi.Control_StatusServer) error {
