+++ engine/server/flusher.go
@@ -1,4 +1,4 @@
-package schema
+package server
 
 import (
 	"net/http"
@@ -36,7 +36,10 @@ func flushAfterNBytes(n int) func(http.Handler) http.Handler {
 				limit:          n,
 				mu:             &sync.Mutex{},
 			}
-			defer flushWriter.Flush()
+			// TODO: double check rm of this defer and explain
+			// saw panic in engine logs (though no test error?)
+			// https://github.com/golang/go/issues/19959
+			// defer flushWriter.Flush()
 
 			next.ServeHTTP(flushWriter, r)
 		})
