+++ engine/server/server.go
@@ -1,38 +1,62 @@
 package server
 
 import (
-	"bytes"
 	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"net"
 	"net/http"
-	"os"
+	"runtime"
+	"runtime/debug"
 	"sync"
 	"time"
 
+	"github.com/99designs/gqlgen/graphql"
+	"github.com/99designs/gqlgen/graphql/handler"
 	"github.com/containerd/containerd/defaults"
 	"github.com/dagger/dagger/analytics"
 	"github.com/dagger/dagger/auth"
 	"github.com/dagger/dagger/core"
 	"github.com/dagger/dagger/core/pipeline"
 	"github.com/dagger/dagger/core/schema"
+	"github.com/dagger/dagger/dagql"
 	"github.com/dagger/dagger/engine"
 	"github.com/dagger/dagger/engine/buildkit"
-	bksession "github.com/moby/buildkit/session"
+	"github.com/dagger/dagger/engine/cache"
+	"github.com/dagger/dagger/engine/client"
+	"github.com/dagger/dagger/tracing"
+	"github.com/moby/buildkit/cache/remotecache"
+	bkgw "github.com/moby/buildkit/frontend/gateway/client"
+	"github.com/moby/buildkit/identity"
+	"github.com/moby/buildkit/session"
 	"github.com/moby/buildkit/util/bklog"
-	bkworker "github.com/moby/buildkit/worker"
+	"github.com/opencontainers/go-digest"
 	"github.com/sirupsen/logrus"
+	"github.com/vektah/gqlparser/v2/gqlerror"
 	"github.com/vito/progrock"
 )
 
 type DaggerServer struct {
 	serverID string
-	bkClient *buildkit.Client
-	worker   bkworker.Worker
 
-	schema      *schema.APIServer
+	clientIDToSecretToken map[string]string
+	connectedClients      int
+	clientIDMu            sync.RWMutex
+
+	// The metadata of client calls.
+	// For the special case of the main client caller, the key is just empty string.
+	// This is never explicitly deleted from; instead it will just be garbage collected
+	// when this server for the session shuts down
+	clientCallContext map[digest.Digest]*core.ClientCallContext
+	clientCallMu      *sync.RWMutex
+
+	// the http endpoints being served (as a map since APIs like shellEndpoint can add more)
+	endpoints  map[string]http.Handler
+	endpointMu *sync.RWMutex
+
+	services *core.Services
+
 	recorder    *progrock.Recorder
 	analytics   analytics.Tracker
 	progCleanup func() error
@@ -40,123 +64,175 @@ type DaggerServer struct {
 	doneCh    chan struct{}
 	closeOnce sync.Once
 
-	connectedClients int
-	clientMu         sync.RWMutex
+	mainClientCallerID        string
+	upstreamCacheExporterCfgs []bkgw.CacheOptionsEntry
+	upstreamCacheExporters    map[string]remotecache.ResolveCacheExporterFunc
 }
 
-func NewDaggerServer(
-	ctx context.Context,
-	bkClient *buildkit.Client,
-	worker bkworker.Worker,
-	caller bksession.Caller,
-	serverID string,
-	secretStore *core.SecretStore,
-	authProvider *auth.RegistryAuthProvider,
-	rootLabels []pipeline.Label,
-	cloudToken string,
-	doNotTrack bool,
-) (*DaggerServer, error) {
-	srv := &DaggerServer{
-		serverID: serverID,
-		bkClient: bkClient,
-		worker:   worker,
-		analytics: analytics.New(analytics.Config{
-			DoNotTrack: doNotTrack || analytics.DoNotTrack(),
-			Labels:     rootLabels,
-			CloudToken: cloudToken,
-		}),
+func (e *BuildkitController) newDaggerServer(ctx context.Context, clientMetadata *engine.ClientMetadata) (*DaggerServer, error) {
+	s := &DaggerServer{
+		serverID: clientMetadata.ServerID,
+
+		clientIDToSecretToken: map[string]string{},
+		clientCallContext:     map[digest.Digest]*core.ClientCallContext{},
+		clientCallMu:          &sync.RWMutex{},
+		endpoints:             map[string]http.Handler{},
+		endpointMu:            &sync.RWMutex{},
+
 		doneCh: make(chan struct{}, 1),
+
+		services: core.NewServices(),
+
+		mainClientCallerID:     clientMetadata.ClientID,
+		upstreamCacheExporters: e.UpstreamCacheExporters,
 	}
 
-	clientConn := caller.Conn()
+	labels := clientMetadata.Labels
+	labels = append(labels, pipeline.EngineLabel(e.EngineName))
+	labels = append(labels, pipeline.LoadServerLabels(engine.Version, runtime.GOOS, runtime.GOARCH, e.cacheManager.ID() != cache.LocalCacheID)...)
+	s.analytics = analytics.New(analytics.Config{
+		DoNotTrack: clientMetadata.DoNotTrack || analytics.DoNotTrack(),
+		Labels:     labels,
+		CloudToken: clientMetadata.CloudToken,
+	})
+
+	getSessionCtx, getSessionCancel := context.WithTimeout(ctx, 10*time.Second)
+	defer getSessionCancel()
+	sessionCaller, err := e.SessionManager.Get(getSessionCtx, clientMetadata.ClientID, false)
+	if err != nil {
+		return nil, fmt.Errorf("get session: %w", err)
+	}
+	clientConn := sessionCaller.Conn()
+
+	// using a new random ID rather than server ID to squash any nefarious attempts to set
+	// a server id that has e.g. ../../.. or similar in it
+	progSockPath := fmt.Sprintf("/run/dagger/server-progrock-%s.sock", identity.NewID())
+
 	progClient := progrock.NewProgressServiceClient(clientConn)
 	progUpdates, err := progClient.WriteUpdates(ctx)
 	if err != nil {
 		return nil, err
 	}
 
-	progWriter, progCleanup, err := buildkit.ProgrockForwarder(bkClient.ProgSockPath, progrock.MultiWriter{
+	progWriter, progCleanup, err := buildkit.ProgrockForwarder(progSockPath, progrock.MultiWriter{
 		progrock.NewRPCWriter(clientConn, progUpdates),
 		buildkit.ProgrockLogrusWriter{},
 	})
 	if err != nil {
 		return nil, err
 	}
-	srv.progCleanup = progCleanup
+	s.progCleanup = progCleanup
 
 	progrockLabels := []*progrock.Label{}
-	for _, label := range rootLabels {
+	for _, label := range labels {
 		progrockLabels = append(progrockLabels, &progrock.Label{
 			Name:  label.Name,
 			Value: label.Value,
 		})
 	}
-	srv.recorder = progrock.NewRecorder(progWriter, progrock.WithLabels(progrockLabels...))
+	s.recorder = progrock.NewRecorder(progWriter, progrock.WithLabels(progrockLabels...))
 
-	// NOTE: context.Background is used because if the provided context is canceled, buildkit can
-	// leave internal progress contexts open and leak goroutines.
-	bkClient.WriteStatusesTo(context.Background(), srv.recorder)
+	secretStore := core.NewSecretStore()
+	authProvider := auth.NewRegistryAuthProvider()
 
-	apiSchema, err := schema.New(ctx, schema.InitializeArgs{
-		BuildkitClient: srv.bkClient,
-		Platform:       srv.worker.Platforms(true)[0],
-		ProgSockPath:   bkClient.ProgSockPath,
-		OCIStore:       srv.worker.ContentStore(),
-		LeaseManager:   srv.worker.LeaseManager(),
-		Secrets:        secretStore,
-		Auth:           authProvider,
+	var cacheImporterCfgs []bkgw.CacheOptionsEntry
+	for _, cacheImportCfg := range clientMetadata.UpstreamCacheImportConfig {
+		_, ok := e.UpstreamCacheImporters[cacheImportCfg.Type]
+		if !ok {
+			return nil, fmt.Errorf("unknown cache importer type %q", cacheImportCfg.Type)
+		}
+		cacheImporterCfgs = append(cacheImporterCfgs, bkgw.CacheOptionsEntry{
+			Type:  cacheImportCfg.Type,
+			Attrs: cacheImportCfg.Attrs,
+		})
+	}
+	for _, cacheExportCfg := range clientMetadata.UpstreamCacheExportConfig {
+		_, ok := e.UpstreamCacheExporters[cacheExportCfg.Type]
+		if !ok {
+			return nil, fmt.Errorf("unknown cache exporter type %q", cacheExportCfg.Type)
+		}
+		s.upstreamCacheExporterCfgs = append(s.upstreamCacheExporterCfgs, bkgw.CacheOptionsEntry{
+			Type:  cacheExportCfg.Type,
+			Attrs: cacheExportCfg.Attrs,
+		})
+	}
+
+	root, err := core.NewRoot(ctx, core.QueryOpts{
+		BuildkitOpts: &buildkit.Opts{
+			Worker:                e.worker,
+			SessionManager:        e.SessionManager,
+			LLBSolver:             e.llbSolver,
+			GenericSolver:         e.genericSolver,
+			SecretStore:           secretStore,
+			AuthProvider:          authProvider,
+			PrivilegedExecEnabled: e.privilegedExecEnabled,
+			UpstreamCacheImports:  cacheImporterCfgs,
+			ProgSockPath:          progSockPath,
+			MainClientCaller:      sessionCaller,
+			MainClientCallerID:    s.mainClientCallerID,
+			DNSConfig:             e.DNSConfig,
+			Frontends:             e.Frontends,
+			RefsMu:                &sync.Mutex{},
+			ExecMetadataMu:        &sync.Mutex{},
+		},
+		ProgrockSocketPath: progSockPath,
+		Services:           s.services,
+		Platform:           core.Platform(e.worker.Platforms(true)[0]),
+		Secrets:            secretStore,
+		OCIStore:           e.worker.ContentStore(),
+		LeaseManager:       e.worker.LeaseManager(),
+		Auth:               authProvider,
+		ClientCallContext:  s.clientCallContext,
+		ClientCallMu:       s.clientCallMu,
+		Endpoints:          s.endpoints,
+		EndpointMu:         s.endpointMu,
+		Recorder:           s.recorder,
 	})
 	if err != nil {
 		return nil, err
 	}
-	srv.schema = apiSchema
-	return srv, nil
-}
 
-func (srv *DaggerServer) LogMetrics(l *logrus.Entry) *logrus.Entry {
-	srv.clientMu.RLock()
-	defer srv.clientMu.RUnlock()
-	return l.WithField(fmt.Sprintf("server-%s-client-count", srv.serverID), srv.connectedClients)
-}
+	dag := dagql.NewServer(root)
 
-func (srv *DaggerServer) Close() {
-	defer srv.closeOnce.Do(func() {
-		close(srv.doneCh)
-	})
+	// stash away the cache so we can share it between other servers
+	root.Cache = dag.Cache
 
-	// mark all groups completed
-	srv.recorder.Complete()
-	// close the recorder so the UI exits
-	srv.recorder.Close()
-	srv.progCleanup()
-	// close the analytics recorder
-	srv.analytics.Close()
-}
+	dag.Around(tracing.AroundFunc)
 
-func (srv *DaggerServer) Wait(ctx context.Context) error {
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	case <-srv.doneCh:
-		return nil
+	coreMod := &schema.CoreMod{Dag: dag}
+	root.DefaultDeps = core.NewModDeps(root, []core.Mod{coreMod})
+	if err := coreMod.Install(ctx, dag); err != nil {
+		return nil, err
 	}
+
+	// the main client caller starts out with the core API loaded
+	s.clientCallContext[""] = &core.ClientCallContext{
+		Deps: root.DefaultDeps,
+		Root: root,
+	}
+
+	return s, nil
 }
 
-func (srv *DaggerServer) ServeClientConn(
+func (s *DaggerServer) ServeClientConn(
 	ctx context.Context,
 	clientMetadata *engine.ClientMetadata,
 	conn net.Conn,
 ) error {
 	bklog.G(ctx).Trace("serve client conn")
 	defer bklog.G(ctx).Trace("done serving client conn")
-	srv.clientMu.Lock()
-	srv.connectedClients++
+	if err := s.VerifyClient(clientMetadata.ClientID, clientMetadata.ClientSecretToken); err != nil {
+		return fmt.Errorf("failed to verify client: %w", err)
+	}
+
+	s.clientIDMu.Lock()
+	s.connectedClients++
 	defer func() {
-		srv.clientMu.Lock()
-		srv.connectedClients--
-		srv.clientMu.Unlock()
+		s.clientIDMu.Lock()
+		s.connectedClients--
+		s.clientIDMu.Unlock()
 	}()
-	srv.clientMu.Unlock()
+	s.clientIDMu.Unlock()
 
 	conn = newLogicalDeadlineConn(splitWriteConn{nopCloserConn{conn}, defaults.DefaultMaxSendMsgSize * 95 / 100})
 	l := &singleConnListener{conn: conn, closeCh: make(chan struct{})}
@@ -165,293 +241,230 @@ func (srv *DaggerServer) ServeClientConn(
 		l.Close()
 	}()
 
-	// NOTE: not sure how inefficient making a new server per-request is, fix if it's meaningful.
-	// Maybe we could dynamically mux in more endpoints for each client or something
-	handler, err := srv.HTTPHandlerForClient(clientMetadata, conn, bklog.G(ctx))
-	if err != nil {
-		return fmt.Errorf("failed to create http handler: %w", err)
-	}
 	httpSrv := http.Server{
-		Handler:           handler,
+		Handler:           s,
 		ReadHeaderTimeout: 30 * time.Second,
+		BaseContext: func(net.Listener) context.Context {
+			ctx := bklog.WithLogger(context.Background(), bklog.G(ctx))
+			ctx = progrock.ToContext(ctx, s.recorder)
+			ctx = engine.ContextWithClientMetadata(ctx, clientMetadata)
+			ctx = analytics.WithContext(ctx, s.analytics)
+			return ctx
+		},
 	}
 	defer httpSrv.Close()
 	return httpSrv.Serve(l)
 }
 
-func (srv *DaggerServer) HTTPHandlerForClient(clientMetadata *engine.ClientMetadata, conn net.Conn, lg *logrus.Entry) (http.Handler, error) {
-	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		defer func() {
-			lg.Trace("handler done")
-		}()
-		req = req.WithContext(bklog.WithLogger(req.Context(), lg))
-		bklog.G(req.Context()).Debugf("http handler for client conn to path %s", req.URL.Path)
-		defer bklog.G(req.Context()).Debugf("http handler for client conn done: %s", clientMetadata.ClientID)
-
-		req = req.WithContext(progrock.ToContext(req.Context(), srv.recorder))
-		req = req.WithContext(engine.ContextWithClientMetadata(req.Context(), clientMetadata))
-		req = req.WithContext(analytics.WithContext(req.Context(), srv.analytics))
-
-		srv.schema.ServeHTTP(w, req)
-	}), nil
-}
-
-// converts a pre-existing net.Conn into a net.Listener that returns the conn and then blocks
-type singleConnListener struct {
-	conn      net.Conn
-	l         sync.Mutex
-	closeCh   chan struct{}
-	closeOnce sync.Once
-}
-
-func (l *singleConnListener) Accept() (net.Conn, error) {
-	l.l.Lock()
-	if l.conn == nil {
-		l.l.Unlock()
-		<-l.closeCh
-		return nil, io.ErrClosedPipe
-	}
-	defer l.l.Unlock()
-
-	c := l.conn
-	l.conn = nil
-	return c, nil
-}
-
-func (l *singleConnListener) Addr() net.Addr {
-	return nil
-}
-
-func (l *singleConnListener) Close() error {
-	l.closeOnce.Do(func() {
-		close(l.closeCh)
-	})
-	return nil
-}
-
-type nopCloserConn struct {
-	net.Conn
-}
-
-func (nopCloserConn) Close() error {
-	return nil
-}
-
-// TODO: could also implement this upstream on:
-// https://github.com/sipsma/buildkit/blob/fa11bf9e57a68e3b5252386fdf44042dd672949a/session/grpchijack/dial.go#L45-L45
-type withDeadlineConn struct {
-	conn          net.Conn
-	readDeadline  time.Time
-	readers       []func()
-	readBuf       *bytes.Buffer
-	readEOF       bool
-	readCond      *sync.Cond
-	writeDeadline time.Time
-	writers       []func()
-	writersL      sync.Mutex
-}
-
-func newLogicalDeadlineConn(inner net.Conn) net.Conn {
-	c := &withDeadlineConn{
-		conn:     inner,
-		readBuf:  new(bytes.Buffer),
-		readCond: sync.NewCond(new(sync.Mutex)),
+func (s *DaggerServer) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	errorOut := func(err error, code int) {
+		bklog.G(ctx).WithError(err).Error("failed to serve request")
+		http.Error(w, err.Error(), code)
 	}
 
-	go func() {
-		for {
-			buf := make([]byte, 32*1024)
-			n, err := inner.Read(buf)
-			if err != nil {
-				c.readCond.L.Lock()
-				c.readEOF = true
-				c.readCond.L.Unlock()
-				c.readCond.Broadcast()
+	clientMetadata, err := engine.ClientMetadataFromContext(ctx)
+	if err != nil {
+		errorOut(err, http.StatusInternalServerError)
+		return
+	}
+
+	callContext, ok := s.ClientCallContext(clientMetadata.ModuleCallerDigest)
+	if !ok {
+		errorOut(fmt.Errorf("client call %s not found", clientMetadata.ModuleCallerDigest), http.StatusInternalServerError)
+		return
+	}
+
+	rec := progrock.FromContext(ctx)
+	if header := r.Header.Get(client.ProgrockParentHeader); header != "" {
+		rec = rec.WithParent(header)
+	} else if callContext.ProgrockParent != "" {
+		rec = rec.WithParent(callContext.ProgrockParent)
+	}
+	ctx = progrock.ToContext(ctx, rec)
+
+	schema, err := callContext.Deps.Schema(ctx)
+	if err != nil {
+		// TODO: technically this is not *always* bad request, should ideally be more specific and differentiate
+		errorOut(err, http.StatusBadRequest)
+		return
+	}
+
+	defer func() {
+		if v := recover(); v != nil {
+			bklog.G(context.TODO()).Errorf("panic serving schema: %v %s", v, string(debug.Stack()))
+			// check whether this is a hijacked connection, if so we can't write any http errors to it
+			_, err := w.Write(nil)
+			if err == http.ErrHijacked {
 				return
 			}
-
-			c.readCond.L.Lock()
-			c.readBuf.Write(buf[0:n])
-			c.readCond.Broadcast()
-			c.readCond.L.Unlock()
+			gqlErr := &gqlerror.Error{
+				Message: "Internal Server Error",
+			}
+			code := http.StatusInternalServerError
+			switch v := v.(type) {
+			case error:
+				gqlErr.Err = v
+				gqlErr.Message = v.Error()
+			case string:
+				gqlErr.Message = v
+			}
+			res := graphql.Response{
+				Errors: gqlerror.List{gqlErr},
+			}
+			bytes, err := json.Marshal(res)
+			if err != nil {
+				panic(err)
+			}
+			http.Error(w, string(bytes), code)
 		}
 	}()
 
-	return c
-}
+	srv := handler.NewDefaultServer(schema)
+	// NB: break glass when needed:
+	// srv.AroundResponses(func(ctx context.Context, next graphql.ResponseHandler) *graphql.Response {
+	// 	res := next(ctx)
+	// 	pl, err := json.Marshal(res)
+	// 	slog.Debug("graphql response", "response", string(pl), "error", err)
+	// 	return res
+	// })
+	mux := http.NewServeMux()
+	mux.Handle("/query", srv)
+	mux.Handle("/shutdown", http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		ctx := req.Context()
+		bklog.G(ctx).Debugf("shutting down client %s", clientMetadata.ClientID)
+		defer bklog.G(ctx).Debugf("shut down client %s", clientMetadata.ClientID)
 
-func (c *withDeadlineConn) Read(b []byte) (n int, err error) {
-	c.readCond.L.Lock()
-
-	if c.readEOF {
-		c.readCond.L.Unlock()
-		return 0, io.EOF
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	if !c.readDeadline.IsZero() {
-		if time.Now().After(c.readDeadline) {
-			c.readCond.L.Unlock()
-			// return early without calling inner Read
-			return 0, os.ErrDeadlineExceeded
+		if len(s.upstreamCacheExporterCfgs) > 0 && clientMetadata.ClientID == s.mainClientCallerID {
+			bklog.G(ctx).Debugf("running cache export for client %s", clientMetadata.ClientID)
+			cacheExporterFuncs := make([]buildkit.ResolveCacheExporterFunc, len(s.upstreamCacheExporterCfgs))
+			for i, cacheExportCfg := range s.upstreamCacheExporterCfgs {
+				cacheExportCfg := cacheExportCfg
+				cacheExporterFuncs[i] = func(ctx context.Context, sessionGroup session.Group) (remotecache.Exporter, error) {
+					exporterFunc, ok := s.upstreamCacheExporters[cacheExportCfg.Type]
+					if !ok {
+						return nil, fmt.Errorf("unknown cache exporter type %q", cacheExportCfg.Type)
+					}
+					return exporterFunc(ctx, sessionGroup, cacheExportCfg.Attrs)
+				}
+			}
+			s.clientCallMu.RLock()
+			bk := s.clientCallContext[""].Root.Buildkit
+			s.clientCallMu.RUnlock()
+			err := bk.UpstreamCacheExport(ctx, cacheExporterFuncs)
+			if err != nil {
+				bklog.G(ctx).WithError(err).Errorf("error running cache export for client %s", clientMetadata.ClientID)
+			}
+			bklog.G(ctx).Debugf("done running cache export for client %s", clientMetadata.ClientID)
 		}
 
-		go func() {
-			dt := time.Until(c.readDeadline)
-			if dt > 0 {
-				time.Sleep(dt)
-			}
-
-			cancel()
-		}()
-	}
-
-	// Keep track of the reader so a future SetReadDeadline can interrupt it.
-	c.readers = append(c.readers, cancel)
-
-	c.readCond.L.Unlock()
-
-	// Start a goroutine for the actual Read operation
-	read := make(chan struct{})
-	var rN int
-	var rerr error
-	go func() {
-		defer close(read)
-
-		c.readCond.L.Lock()
-		defer c.readCond.L.Unlock()
-
-		for ctx.Err() == nil {
-			if c.readEOF {
-				rerr = io.EOF
-				break
-			}
-
-			n, _ := c.readBuf.Read(b) // ignore EOF here
-			if n > 0 {
-				rN = n
-				break
-			}
-
-			c.readCond.Wait()
+		// TODO:(XXX) double-check this is still correct, don't break services across modules
+		if err := s.services.StopClientServices(ctx, clientMetadata); err != nil {
+			bklog.G(ctx).WithError(err).Error("failed to shutdown")
 		}
-	}()
-
-	// Wait for either Read to complete or the timeout
-	select {
-	case <-read:
-		return rN, rerr
-	case <-ctx.Done():
-		return 0, os.ErrDeadlineExceeded
+	}))
+	s.endpointMu.RLock()
+	for path, handler := range s.endpoints {
+		mux.Handle(path, handler)
 	}
+	s.endpointMu.RUnlock()
+
+	r = r.WithContext(ctx)
+
+	var handler http.Handler = mux
+	handler = flushAfterNBytes(buildkit.MaxFileContentsChunkSize)(handler)
+	handler.ServeHTTP(w, r)
 }
 
-func (c *withDeadlineConn) Write(b []byte) (n int, err error) {
-	c.writersL.Lock()
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	if !c.writeDeadline.IsZero() {
-		if time.Now().After(c.writeDeadline) {
-			c.writersL.Unlock()
-			// return early without calling inner Write
-			return 0, os.ErrDeadlineExceeded
+func (s *DaggerServer) RegisterClient(clientID, clientHostname, secretToken string) error {
+	s.clientIDMu.Lock()
+	defer s.clientIDMu.Unlock()
+	existingToken, ok := s.clientIDToSecretToken[clientID]
+	if ok {
+		if existingToken != secretToken {
+			return fmt.Errorf("client ID %q already registered with different secret token", clientID)
 		}
-
-		go func() {
-			dt := time.Until(c.writeDeadline)
-			if dt > 0 {
-				time.Sleep(dt)
-			}
-
-			cancel()
-		}()
+		return nil
 	}
-
-	// Keep track of the writer so a future SetWriteDeadline can interrupt it.
-	c.writers = append(c.writers, cancel)
-	c.writersL.Unlock()
-
-	// Start a goroutine for the actual Write operation
-	write := make(chan int, 1)
-	go func() {
-		n, err = c.conn.Write(b)
-		write <- 0
-	}()
-
-	// Wait for either Write to complete or the timeout
-	select {
-	case <-write:
-		return n, err
-	case <-ctx.Done():
-		return 0, os.ErrDeadlineExceeded
-	}
-}
-
-func (c *withDeadlineConn) Close() error {
-	return c.conn.Close()
-}
-
-func (c *withDeadlineConn) LocalAddr() net.Addr {
-	return c.conn.LocalAddr()
-}
-
-func (c *withDeadlineConn) RemoteAddr() net.Addr {
-	return c.conn.RemoteAddr()
-}
-
-func (c *withDeadlineConn) SetDeadline(t time.Time) error {
-	return errors.Join(
-		c.SetReadDeadline(t),
-		c.SetWriteDeadline(t),
-	)
-}
-
-func (c *withDeadlineConn) SetReadDeadline(t time.Time) error {
-	c.readCond.L.Lock()
-	c.readDeadline = t
-	readers := c.readers
-	c.readCond.L.Unlock()
-
-	if len(readers) > 0 && !t.IsZero() {
-		go func() {
-			dt := time.Until(c.readDeadline)
-			if dt > 0 {
-				time.Sleep(dt)
-			}
-
-			for _, cancel := range readers {
-				cancel()
-			}
-		}()
-	}
+	s.clientIDToSecretToken[clientID] = secretToken
+	// NOTE: we purposely don't delete the secret token, it should never be reused and will be released
+	// from memory once the dagger server instance corresponding to this buildkit client shuts down.
+	// Deleting it would make it easier to create race conditions around using the client's session
+	// before it is fully closed.
 
 	return nil
 }
 
-func (c *withDeadlineConn) SetWriteDeadline(t time.Time) error {
-	c.writersL.Lock()
-	c.writeDeadline = t
-	writers := c.writers
-	c.writersL.Unlock()
-
-	if len(writers) > 0 && !t.IsZero() {
-		go func() {
-			dt := time.Until(c.writeDeadline)
-			if dt > 0 {
-				time.Sleep(dt)
-			}
-
-			for _, cancel := range writers {
-				cancel()
-			}
-		}()
+func (s *DaggerServer) VerifyClient(clientID, secretToken string) error {
+	s.clientIDMu.RLock()
+	defer s.clientIDMu.RUnlock()
+	existingToken, ok := s.clientIDToSecretToken[clientID]
+	if !ok {
+		return fmt.Errorf("client ID %q not registered", clientID)
 	}
-
+	if existingToken != secretToken {
+		return fmt.Errorf("client ID %q registered with different secret token", clientID)
+	}
 	return nil
+}
+
+func (s *DaggerServer) ClientCallContext(clientDigest digest.Digest) (*core.ClientCallContext, bool) {
+	s.clientCallMu.RLock()
+	defer s.clientCallMu.RUnlock()
+	ctx, ok := s.clientCallContext[clientDigest]
+	return ctx, ok
+}
+
+func (s *DaggerServer) CurrentServedDeps(ctx context.Context) (*core.ModDeps, error) {
+	clientMetadata, err := engine.ClientMetadataFromContext(ctx)
+	if err != nil {
+		return nil, err
+	}
+	callCtx, ok := s.ClientCallContext(clientMetadata.ModuleCallerDigest)
+	if !ok {
+		return nil, fmt.Errorf("client call %s not found", clientMetadata.ModuleCallerDigest)
+	}
+	return callCtx.Deps, nil
+}
+
+func (s *DaggerServer) LogMetrics(l *logrus.Entry) *logrus.Entry {
+	s.clientIDMu.RLock()
+	defer s.clientIDMu.RUnlock()
+	return l.WithField(fmt.Sprintf("server-%s-client-count", s.serverID), s.connectedClients)
+}
+
+func (s *DaggerServer) Close() error {
+	defer s.closeOnce.Do(func() {
+		close(s.doneCh)
+	})
+
+	var err error
+
+	s.clientCallMu.RLock()
+	for _, callCtx := range s.clientCallContext {
+		// TODO:(XXX) should we clean up services here too?
+		err = errors.Join(err, callCtx.Root.Buildkit.Close())
+	}
+	s.clientCallMu.RUnlock()
+
+	// mark all groups completed
+	s.recorder.Complete()
+	// close the recorder so the UI exits
+	err = errors.Join(err, s.recorder.Close())
+	err = errors.Join(err, s.progCleanup())
+	// close the analytics recorder
+	err = errors.Join(err, s.analytics.Close())
+
+	return err
+}
+
+func (s *DaggerServer) Wait(ctx context.Context) error {
+	select {
+	case <-ctx.Done():
+		return ctx.Err()
+	case <-s.doneCh:
+		return nil
+	}
 }
 
 type splitWriteConn struct {
